requirements.txt
langchain
langchain-community
langchain-huggingface
pypdf
sentence-transformers
chromadb
python-dotenv